Expand Azure Pipelines and Integrate GitHub Actions

+++ testing/adios2/engine/bp/TestBPWriteReadLocalVariables.cpp
@@ -1864,8 +1864,9 @@ TEST_F(BPWriteReadLocalVariables, ADIOS2BPWriteReadLocal2DChangeCount)
 #endif
 
     const size_t Nx0 =
-        static_cast<size_t>(std::pow(2 - static_cast<size_t>(mpiRank), 2)) + 1;
+        static_cast<size_t>(std::pow(2 - mpiRank, 2)) + 1;
     const size_t Nx1 = 5;
+
     // data for block 0 and block 1 per rank
     std::vector<float> data(Nx0 * Nx1);
     const float startBlock = static_cast<float>(3 * mpiRank + 1);
